Function for handling feature selection

Function for handling feature selection

Usage

handleFeatureSelection(
  data,
  pathToResults,
  topK,
  prevalenceCutOff,
  targetCohortId,
  runZTests,
  runLogitTests,
  createOutputFiles = FALSE
)

Arguments

data

Data list object

pathToResults

Path to the results folder, can be project's working directory

topK

numeric > if set, keeps this number of features in the analysis. Maximum number of features exported.

prevalenceCutOff

numeric > if set, removes all of the concepts which are not present (in target) more than prevalenceCutOff times

targetCohortId

Target cohort id

runZTests

boolean for Z-tests

runLogitTests

boolean for logit-tests

createOutputFiles

Boolean for creating output files, the default value is TRUE