Match Pattern in Column Names

This function searches for specified patterns within column names provided. It takes one or more patterns as input and returns a list of matches. Each pattern is applied to the list of column names, and if a pattern does not match any column, an error is thrown.

Usage

patterns(..., cols = character(0L))

Arguments

...

Character, patterns to search for within the column names. You can provide multiple patterns as separate arguments.

cols

A character vector of column names against which the patterns will be matched. Defaults to an empty character vector, and it expects the actual column names to be provided to function properly.

Value

A list where each element corresponds to a pattern provided in `...`. Each list element contains the indices of the column names in `cols` that match the respective pattern. If a pattern does not match any column names, the function will stop and throw an error indicating which pattern(s) did not match.